backend/pyproject.toml
[tool.poetry]
name = "dev-challenge-backend"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
package-mode=false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
django-rest-framework = "^0.1.0"
requests = "^2.32.3"
django-rq = "^2.10.2"
loguru = "^0.7.2"

[tool.poetry.group.dev.dependencies]
django-dotenv = "^1.4.2"
jupyter = "^1.1.1"

[tool.poetry.group.windows]
optional = true

[tool.poetry.group.windows.dependencies]
psycopg2 = "^2.9.6"

[tool.poetry.group.linux]
optional = true

[tool.poetry.group.linux.dependencies]
psycopg2-binary = "^2.9.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
